+++ rest_framework_json_api/parsers.py
@@ -139,6 +139,17 @@ class JSONParser(parsers.JSONParser):
         if not data.get('id') and request.method in ('PATCH', 'PUT'):
             raise ParseError("The resource identifier object must contain an 'id' member")
 
+        if request.method in ('PATCH', 'PUT'):
+            lookup_url_kwarg = view.lookup_url_kwarg or view.lookup_field
+            if str(data.get('id')) != str(view.kwargs[lookup_url_kwarg]):
+                raise exceptions.Conflict(
+                    "The resource object's id ({data_id}) does not match url's "
+                    "lookup id ({url_id})".format(
+                        data_id=data.get('id'),
+                        url_id=view.kwargs[view.lookup_field]
+                    )
+                )
+
         # Construct the return data
         serializer_class = getattr(view, 'serializer_class', None)
         parsed_data = {'id': data.get('id')} if 'id' in data else {}